Rename method_types.py -> method_lists.py
